refacroted & added comments for maintainers
