+++ patches/minecraft/net/minecraft/client/renderer/chunk/RenderChunk.java.patch
@@ -1,6 +1,19 @@
 --- ../src-base/minecraft/net/minecraft/client/renderer/chunk/RenderChunk.java
 +++ ../src-work/minecraft/net/minecraft/client/renderer/chunk/RenderChunk.java
-@@ -162,7 +162,7 @@
+@@ -142,10 +142,12 @@
+         {
+             p_178581_4_.func_178540_f().unlock();
+         }
++        if (net.minecraftforge.client.ForgeHooksClient.onRebuildChunkEvent(field_178589_e, field_189564_r, p_178581_4_, compiledchunk,p_178581_1_,p_178581_2_,p_178581_3_).isCanceled()) return;
+ 
+         VisGraph lvt_9_1_ = new VisGraph();
+         HashSet lvt_10_1_ = Sets.newHashSet();
+ 
++        final net.minecraftforge.client.event.RenderChunkEvent renderEvent = net.minecraftforge.client.ForgeHooksClient.onRenderChunkEvent(field_178589_e, field_189564_r, p_178581_4_, compiledchunk, BlockPos.func_177975_b(blockpos, blockpos1),Minecraft.func_71410_x().func_175602_ab(),p_178581_1_,p_178581_2_,p_178581_3_);
+         if (!this.field_189564_r.func_72806_N())
+         {
+             ++field_178592_a;
+@@ -162,7 +164,7 @@
                      lvt_9_1_.func_178606_a(blockpos$mutableblockpos);
                  }
  
@@ -9,7 +22,7 @@
                  {
                      TileEntity tileentity = this.field_189564_r.func_190300_a(blockpos$mutableblockpos, Chunk.EnumCreateEntityType.CHECK);
  
-@@ -172,17 +172,19 @@
+@@ -172,17 +174,20 @@
  
                          if (tileentityspecialrenderer != null)
                          {
@@ -25,13 +38,14 @@
                  }
  
 -                BlockRenderLayer blockrenderlayer1 = block.func_180664_k();
++                if (renderEvent.isCanceled()) continue;
 +                for(BlockRenderLayer blockrenderlayer1 : BlockRenderLayer.values()) {
 +                     if(!block.canRenderInLayer(iblockstate, blockrenderlayer1)) continue;
 +                     net.minecraftforge.client.ForgeHooksClient.setRenderLayer(blockrenderlayer1);
                  int j = blockrenderlayer1.ordinal();
  
                  if (block.func_176223_P().func_185901_i() != EnumBlockRenderType.INVISIBLE)
-@@ -197,6 +199,8 @@
+@@ -197,10 +202,13 @@
  
                      aboolean[j] |= blockrendererdispatcher.func_175018_a(iblockstate, blockpos$mutableblockpos, this.field_189564_r, bufferbuilder);
                  }
@@ -40,7 +54,12 @@
              }
  
              for (BlockRenderLayer blockrenderlayer : BlockRenderLayer.values())
-@@ -278,7 +282,9 @@
+             {
++                if (renderEvent.isCanceled()) continue;
+                 if (aboolean[blockrenderlayer.ordinal()])
+                 {
+                     compiledchunk.func_178486_a(blockrenderlayer);
+@@ -278,7 +286,9 @@
      private void func_189563_q()
      {
          int i = 1;
@@ -51,7 +70,7 @@
      }
  
      @Nullable
-@@ -427,6 +433,26 @@
+@@ -427,6 +437,26 @@
          return this.field_178593_n && this.field_188284_q;
      }
