+++ src/main/java/net/minecraftforge/client/event/RebuildChunkEvent.java
@@ -18,29 +18,43 @@
  */
 package net.minecraftforge.client.event;
 
+import net.minecraft.client.renderer.BlockRendererDispatcher;
+import net.minecraft.client.renderer.BufferBuilder;
 import net.minecraft.client.renderer.RenderGlobal;
 import net.minecraft.client.renderer.chunk.ChunkCompileTaskGenerator;
 import net.minecraft.client.renderer.chunk.CompiledChunk;
+import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
+import net.minecraft.util.BlockRenderLayer;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.BlockPos.MutableBlockPos;
 import net.minecraft.world.ChunkCache;
 import net.minecraftforge.fml.common.eventhandler.Cancelable;
 import net.minecraftforge.fml.common.eventhandler.Event;
 
+/**
+ * Called when a {@link net.minecraft.client.renderer.chunk.RenderChunk#rebuildChunk RenderChunk.rebuildChunk} is called. 
+ * This event is fired on {@link net.minecraftforge.common.MinecraftForge#EVENT_BUS}
+ * right after the CompiledChunk is generated and before any rebuilding is done.
+ * Canceling this event prevents all Blocks and Tile Entities from being rebuild (and therefore rendered)
+ */
 @Cancelable
 public class RebuildChunkEvent extends Event {
 
     private final RenderGlobal              context;
-    private final CompiledChunk             compiledChunk;
-    private final ChunkCompileTaskGenerator generator;
     private final ChunkCache                worldView;
+    private final ChunkCompileTaskGenerator generator;
+    private final CompiledChunk             compiledChunk;
+    private final MutableBlockPos           position;
     private final float                     x;
     private final float                     y;
     private final float                     z;
 
-    public RebuildChunkEvent(final RenderGlobal renderGlobal, final ChunkCache worldView, final ChunkCompileTaskGenerator generator, final CompiledChunk compiledChunk, final float x, final float y, final float z) {
+    public RebuildChunkEvent(final RenderGlobal renderGlobal, final ChunkCache worldView, final ChunkCompileTaskGenerator generator, final CompiledChunk compiledChunk, final MutableBlockPos position, final float x, final float y, final float z) {
         this.context = renderGlobal;
         this.worldView = worldView;
         this.generator = generator;
         this.compiledChunk = compiledChunk;
+        this.position = position;
         this.x = x;
         this.y = y;
         this.z = z;
@@ -60,6 +74,118 @@ public class RebuildChunkEvent extends Event {
 
     public CompiledChunk getCompiledChunk() {
         return this.compiledChunk;
+    }
+    
+    public MutableBlockPos getPosition() {
+        return position;
+    }
+    
+    public float getX() {
+        return x;
+    }
+    
+    public float getY() {
+        return y;
+    }
+    
+    public float getZ() {
+        return z;
+    }
+    
+    /**
+     * Called when a {@link net.minecraft.client.renderer.chunk.RenderChunk#rebuildChunk RenderChunk.rebuildChunk} is called. 
+     * This event is fired on {@link net.minecraftforge.common.MinecraftForge#EVENT_BUS}
+     * right before any rebuilding is done.
+     * Canceling this event prevents all Blocks from being rebuild (and therefore rendered).
+     * TileEntities will still be rendered if this event is cancelled
+     */
+    @Cancelable
+    public static class RebuildChunkBlocksEvent extends Event {
+
+        private final RenderGlobal              context;
+        private final ChunkCache                worldView;
+        private final ChunkCompileTaskGenerator generator;
+        private final CompiledChunk             compiledChunk;
+        private final Iterable<MutableBlockPos> chunkBlockPositions;
+        private final BlockRendererDispatcher   blockRendererDispatcher;
+        private final MutableBlockPos           position;
+        private final float                     x;
+        private final float                     y;
+        private final float                     z;
+
+        public RebuildChunkBlocksEvent(final RenderGlobal renderGlobal, final ChunkCache worldView, final ChunkCompileTaskGenerator generator, final CompiledChunk compiledChunk, final Iterable<MutableBlockPos> chunkBlockPositions, final BlockRendererDispatcher blockRendererDispatcher, final MutableBlockPos position, final float x, final float y, final float z) {
+            this.context = renderGlobal;
+            this.worldView = worldView;
+            this.generator = generator;
+            this.compiledChunk = compiledChunk;
+            this.chunkBlockPositions = chunkBlockPositions;
+            this.blockRendererDispatcher = blockRendererDispatcher;
+            this.position = position;
+            this.x = x;
+            this.y = y;
+            this.z = z;
+        }
+
+        public RenderGlobal getContext() {
+            return this.context;
+        }
+
+        public ChunkCache getWorldView() {
+            return this.worldView;
+        }
+
+        public ChunkCompileTaskGenerator getGenerator() {
+            return this.generator;
+        }
+
+        public CompiledChunk getCompiledChunk() {
+            return this.compiledChunk;
+        }
+
+        public Iterable<MutableBlockPos> getChunkBlockPositions() {
+            return this.chunkBlockPositions;
+        }
+
+        public BlockRendererDispatcher getBlockRendererDispatcher() {
+            return this.blockRendererDispatcher;
+        }
+        
+        public MutableBlockPos getPosition() {
+            return position;
+        }
+        
+        public float getX() {
+            return x;
+        }
+        
+        public float getY() {
+            return y;
+        }
+        
+        public float getZ() {
+            return z;
+        }
+
+        public BufferBuilder getBufferBuilder(final BlockRenderLayer blockRenderLayer) {
+            return this.getGenerator().getRegionRenderCacheBuilder().getWorldRendererByLayer(blockRenderLayer);
+        }
+
+        public void preRenderBlocks(final BufferBuilder bufferBuilderIn, final BlockPos pos) {
+            bufferBuilderIn.begin(7, DefaultVertexFormats.BLOCK);
+            bufferBuilderIn.setTranslation((double) (-pos.getX()), (double) (-pos.getY()), (double) (-pos.getZ()));
+        }
+        
+        public void postRenderBlocks(BlockRenderLayer layer, float x, float y, float z, BufferBuilder bufferBuilderIn, CompiledChunk compiledChunkIn)
+        {
+            if (layer == BlockRenderLayer.TRANSLUCENT && !compiledChunkIn.isLayerEmpty(layer))
+            {
+                bufferBuilderIn.sortVertexData(x, y, z);
+                compiledChunkIn.setState(bufferBuilderIn.getVertexState());
+            }
+
+            bufferBuilderIn.finishDrawing();
+        }
+        
     }
 
 }
